Implemented operation choice
Implemented operation choice from the start of the program, wrapped in a do while loop. This will be developed later as I begin to define class systems for each arithmetic operation.

Also removed the redundant userChoice typedef as it served no real purpose; this change may apply to the subsequent typedefs as development continues.