Added additional program output to inform user
Added program feedback to user when a process is about to begin or has concluded. This lets the user know which stage the operation is in.

Additional changes in this commit:
- Altered files with iterative loops, replacing iterative int types to auto for better readability.
- Moved selection screen output back into do while loops; this was to stop the user from having to scroll back up to see the options in the event they decided not to close the program when asked.
- Formatted vecAdd and matMulti selection screens to be in line with opChoice screens

For future commit(s):
Currently experimenting with true 2D vectors as the current implementation feigns 2D matrices. The matMulti operation uses conSize * conSize to imitate the 2D row / column component (see line 10 (file: matMultiCore.cpp)). This merely initialises vectors with double the user selected size.