Fully Impl. OperationEventHandler class
OperationEventHandler class has been fully implemented and integrated with all events throughout the program. With this, I plan to refactor the code to be less reliant on switch statements, and push more towards using a hashmap.

With that said, I plan to have OperationResultHandler no longer use .csv files; results should be stored on an SQL database. This will be implemented using SQLite as it's self-contained and serverless.