Prog. made on GPU matMulti - hit blocker
Blocker relates to the GPU kernel code not executing matrix multiplication past the pre-squared container size. E.g. threads will run until index 1023 and ignore the rest of the vector.

This doesn't seem to be a problem in the functional version of this matMulti, so I'll need to do some digging to figure out what's causing this issue in my code.